pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-authkit"
version = "0.1.0"
description = "Reusable Django authentication, user, role, permission, and social auth APIs."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "django-authkit maintainers" }
]
maintainers = [
  { name = "django-authkit maintainers" }
]
keywords = [
  "django",
  "authentication",
  "authorization",
  "django-rest-framework",
  "openapi",
  "swagger",
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
  "Framework :: Django :: 5.2",
  "Framework :: Django :: 6.0",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "Django>=4.2",
  "djangorestframework>=3.15",
  "djangorestframework-simplejwt>=5.3",
  "drf-spectacular>=0.27",
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "build>=1.2",
  "isort>=5.13",
  "mypy>=1.8",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-django>=4.8",
  "ruff>=0.5",
  "twine>=5.0",
]
social = [
  "google-auth>=2.29",
  "PyJWT>=2.8",
  "cryptography>=42.0",
  "requests>=2.31",
]
docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
]

[project.urls]
Homepage = "https://github.com/django-authkit/django-authkit"
Documentation = "https://github.com/django-authkit/django-authkit/tree/main/docs"
Repository = "https://github.com/django-authkit/django-authkit"
Issues = "https://github.com/django-authkit/django-authkit/issues"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["authkit*"]
exclude = [
  "tests*",
  "docs*",
  "example_django_app*",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
py_version = 312
src_paths = ["src", "tests", "example_django_app"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests", "example_django_app"]
exclude = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".venv",
  "build",
  "dist",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "DJ"]
ignore = [
  "DJ008",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["DJ"]

[tool.mypy]
python_version = "3.12"
plugins = []
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
files = ["src/authkit"]
exclude = [
  "build/",
  "dist/",
  "example_django_app/",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
DJANGO_SETTINGS_MODULE = "tests.settings"
django_find_project = false
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = [
  "--strict-config",
  "--strict-markers",
  "-ra",
]

[tool.coverage.run]
branch = true
source = ["authkit"]
omit = [
  "src/authkit/migrations/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 85
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
]

[tool.coverage.html]
directory = "htmlcov"
